+++ app/forms/conversion/create_project_form.rb
@@ -80,7 +80,7 @@ class Conversion::CreateProjectForm < CreateProjectForm
       urn: urn,
       incoming_trust_ukprn: incoming_trust_ukprn,
       establishment_sharepoint_link: establishment_sharepoint_link,
-      trust_sharepoint_link: trust_sharepoint_link,
+      incoming_trust_sharepoint_link: incoming_trust_sharepoint_link,
       advisory_board_conditions: advisory_board_conditions,
       conversion_date: provisional_conversion_date,
       advisory_board_date: advisory_board_date,