+++ spec/factories/conversion/create_project_form_factory.rb
@@ -5,7 +5,7 @@ FactoryBot.define do
     provisional_conversion_date { {3 => 1, 2 => 1, 1 => 2030} }
     advisory_board_date { {3 => 1, 2 => 10, 1 => 2022} }
     establishment_sharepoint_link { "https://educationgovuk-my.sharepoint.com/establishment-folder" }
-    trust_sharepoint_link { "https://educationgovuk-my.sharepoint.com/trust-folder" }
+    incoming_trust_sharepoint_link { "https://educationgovuk-my.sharepoint.com/trust-folder" }
     user { association :user, :regional_delivery_officer }
     handover_note_body { "Handover notes" }
     directive_academy_order { false }